requirements.txt
ollama>=0.4.0
requests>=2.32.5
charset-normalizer>=3.4.4
urllib3>=2.6.3
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
opencv-python>=4.13.0
python-dotenv>=1.0.0